requirements.txt
pydantic-settings==2.5.2
pydantic==2.9.1
mypy==1.11.2
ruff==0.6.7

datarobot==3.5.2
datarobot-predict==1.9.2

openai~=1.31.2
pandas~=2.2.2

pytest==8.0.1

pulumi==3.133.0
pulumi-datarobot==0.1.44

datarobotx-idp @ git+https://github.com/datarobot-community/datarobotx-idp.git